% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brand.R
\name{compare_branding_files}
\alias{compare_branding_files}
\title{Helper function to compare local and remote branding files using MD5 hashes}
\usage{
compare_branding_files(local_file, remote_file)
}
\arguments{
\item{local_file}{Path to the local branding file}

\item{remote_file}{Path to the remote branding file}
}
\value{
TRUE if files are identical, FALSE otherwise
}
\description{
Helper function to compare local and remote branding files using MD5 hashes
}
