% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brand.R
\name{download_branding_file}
\alias{download_branding_file}
\title{Helper function for downloading the branding file}
\usage{
download_branding_file(remote_file, auth_token)
}
\arguments{
\item{remote_file}{URL of the remote branding file}

\item{auth_token}{Authentication token for accessing private repositories}
}
\value{
Path to the temporary file where the remote branding file is downloaded
}
\description{
Helper function for downloading the branding file
}
