% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_template.R
\name{get_example}
\alias{get_example}
\title{Copy Example App Files from inst/examples to inst/app}
\usage{
get_example(example_name)
}
\arguments{
\item{example_name}{Name of the example folder under inst/examples}
}
\description{
Copy Example App Files from inst/examples to inst/app
}
