% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example1.R
\name{get_shiny_example1}
\alias{get_shiny_example1}
\title{Copy Example App File to User Directory}
\usage{
get_shiny_example1()
}
\description{
This function creates the folders \code{inst/app} in the user's working directory and copies the example file \code{link_plots.R} into that folder.
}
