% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_template.R
\name{get_template2}
\alias{get_template2}
\title{Copy Example App File to User Directory}
\usage{
get_template2()
}
\description{
This function creates the folders \code{inst/app} in the user's working directory and copies the example file \code{link_plots.R} into that folder.
}
