% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brand.R
\name{update_branding_file}
\alias{update_branding_file}
\title{Helper function to update the local branding file}
\usage{
update_branding_file(local_file, remote_file, backup = FALSE)
}
\arguments{
\item{local_file}{Path to the local branding file}

\item{remote_file}{Path to the remote branding file}

\item{backup}{Logical indicating whether to create a backup of the local file}
}
\value{
NULL. Called for its side effects: updating the local branding file and possibly
creating a backup file
}
\description{
Helper function to update the local branding file
}
