% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand_init.R
\name{brand_init}
\alias{brand_init}
\title{Initialize branding configuration}
\usage{
brand_init()
}
\value{
NULL. Called for its side effects: creating \code{rbranding_config.yml} and \verb{_brand.yml}
}
\description{
\code{brand_init} initializes the branding configuration by creating two files:
\itemize{
\item \code{rbranding_config.yml}: contains remote and local file paths to brand files
\item \verb{_brand.yml}: a placeholder branding file
It is intended to be run once. Use \code{get_brand()} to download/update the brand
file.
}
}
\examples{
brand_init()

# Clean up
file.remove("rbranding_config.yml", "_brand.yml")
}
