% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_repo.R
\name{brand_init}
\alias{brand_init}
\title{Initialize Branding Config}
\usage{
brand_init(get_default_brand = TRUE)
}
\arguments{
\item{get_default_brand}{Logical. If TRUE, calls get_brand() to download the latest branding file after initialization. Default is TRUE.}
}
\value{
No return value. Side effects: creates config.yml and _brand.yml.
}
\description{
Initializes the branding by creating a config.yml file with the remote and local file paths, and a placeholder _brand.yml file.
}
\details{
This function is intended to be run once to set up the configuration for branding file management.
}
\examples{
brand_init(get_default_brand = FALSE)
}
