% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brand.R
\name{get_brand}
\alias{get_brand}
\title{Download the latest branding file}
\usage{
get_brand(
  brand_url = NULL,
  host_url = NULL,
  local_file = NULL,
  config_file = "rbranding_config.yml",
  backup_folder = "."
)
}
\arguments{
\item{brand_url}{Optional URL. Points to the remote brand file. If \code{NULL}, the value
in the configuration file will be used.}

\item{host_url}{Optional URL. Points to the remote host. If \code{NULL}, the value
in the configuration file will be used.}

\item{local_file}{Optional string. Path to the local branding file. If \code{NULL},
the value in the configuration file will be used.}

\item{config_file}{Path to the configuration file. Default is \code{rbranding_config.yml}.}

\item{backup_folder}{Folder where the backup file should be saved, if needed.
Defaults to current working directory.}
}
\value{
NULL. Called for its side effects: updating \verb{_brand.yml} and possibly creating \code{bak_brand.yml}
}
\description{
\code{get_brand} downloads the latest \verb{_brand.yml} file from the remote host
and URL specified in \code{rbranding_config.yml}. If the local \verb{_brand.yml}
file does not exist, it will be created. If the local file is different
from the remote file, the function will save the contents to
\code{bak_brand.yml} (as backup) and overwrite the local file with the contents
of the remote file. When the function is run interactively (e.g., in
RStudio console), the user is instead prompted to choose whether to
overwrite the file and whether or not to create the backup.
}
\examples{
if (interactive()) {
 # Initialize config and local brand file
 brand_init()

 # Update local brand file if needed
 get_brand()

 # Cleanup
 file.remove("rbranding_config.yml", "_brand.yml", "bak_brand.yml")
}
}
