% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_template.R
\name{get_template}
\alias{get_template}
\title{Get Shiny Templates
Copy Example App Files from inst/examples to inst/app}
\usage{
get_template(example_name)
}
\arguments{
\item{example_name}{Name of the example folder under inst/examples}
}
\description{
Get Shiny Templates
Copy Example App Files from inst/examples to inst/app
}
